% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_anova.R
\name{anova_stat}
\alias{anova_stat}
\title{Comparison of raman intensity with multitreatment sample}
\usage{
anova_stat(data, t.name = NULL, var.name = NULL)
}
\arguments{
\item{t.name:}{columnname of multiple treatment}

\item{var.name:}{variable name (Raman shift in ex_data)}
}
\value{
statiscal comparison of raman intensity distribution by raman shift (more than 3 treatment)
}
\description{
Compare intensity distribution by treatment using anova according to variable
}
\examples{
ex=anova_stat(mt,t.name = "Treatment", var.name = "variable")

ggplot(ex)+
geom_line(aes(x=variable, y=F.value),col="red")+
geom_line(aes(x=variable, y=P.value*10), col="royalblue")+
theme_bw()
}
