% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peak_detect.R
\name{peak_detect}
\alias{peak_detect}
\title{peak detection with statistical difference}
\usage{
peak_detect(data = data, g.name = NULL, bin = 2, sig.level = 0.05)
}
\arguments{
\item{g.name:}{columnname of groupname}

\item{bin:}{allow peak gap}

\item{sig.level:}{criteria p.value to identify peak with significant difference}
}
\value{
Sp: Compared sp with control, if data not include control or compare add any column (ex: data$fac="none")

PS: peak start variable

PE: peak start variable

no: temporal peak id while peak detection

gap: count variable between PS and PE
}
\description{
statiscal comparison of raman intensity distribution by raman shift (more than 3 treatment)
}
\examples{
ex=anova_stat(mt,t.name = "Treatment", var.name = "variable")
ex$Trt="none"

t=peak_detect(ex,g.name = "Trt",bin = 2, sig.level = 0.01)
t2=dcast(mt, Sam+Sp+Treatment~variable,mean) \%>\%  melt(id.var=c("Sam","Sp","Treatment"))
write.csv(t2, file = "test.csv",row.names = F)
t2=read.csv("test.csv")
ggplot()+
geom_rect(data=t, inherit.aes = F,aes(xmin=PS, xmax=PE),ymin=-Inf, ymax=Inf,fill="grey90" ,alpha=1)+
geom_line(data=t2,aes(x=variable, y=value, col=Treatment))
ggplot()+

}
