% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tt_ctr.R
\name{ctr.tt}
\alias{ctr.tt}
\title{Comparison of raman spectrum intensity with control}
\usage{
ctr.tt(data = data, g.name, var.name, ctr.name = "a0")
}
\arguments{
\item{g.name:}{group name of comparison target}

\item{var.name:}{variable name (Raman shift)}

\item{ctr.name:}{control sp in group}
}
\value{
statiscal comparison of raman intensity distributionby raman shift
}
\description{
Compare intensity between sample and control according to raman shift
}
\examples{
ft=ctr.tt(data = ex_rm,g.name ="Sp",var.name = "variable", ctr.name = "a0")
head(ft)

ggplot(ft, aes(x=variable, y=p,group=Comp.sp))+
geom_line()+
facet_wrap(.~Comp.sp, ncol = 1)
}
